extract DatabaseEntry to its own file
